+++ CHANGELOG.md
@@ -17,6 +17,7 @@
    - Support for reverse bindings (e.g., `Codec[Point3D].as[Int :: Int :: Int :: HNil]`)
    - Support for singleton bindings without having to first lift `Codec[A]` to `Codec[A :: HNil]` (e.g., `case class Foo(x: Int); uint8.as[Foo]`)
  - Added `optional` combinator
+ - Added `withDefault` and `withDefaultValue` combinators
 
 1.2.0
 =====